api/omqueryinternal.cc,api/omstem.cc, backends/inmemory/inmemory_database.h, backends/multi/multi_termlist.h,backends/quartz/, extra/queryparser.yy,matcher/,net/tcpserver.cc,tests/api_anydb.cc, tests/api_db.cc,tests/quartztest.cc,testsuite/backendmanager.cc, testsuite/btreecheck.cc,testsuite/testsuite.cc: Changed C style casts to C++ style.  The syntax is ugly, but they do make the intent clearer which is a good thing.
git-svn-id: file:///home/xapian-svn/xapian/trunk@5694 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/matcher/expandweight.cc
@@ -3,7 +3,7 @@
  * ----START-LICENCE----
  * Copyright 1999,2000,2001 BrightStation PLC
  * Copyright 2002 Ananova Ltd
- * Copyright 2003 Olly Betts
+ * Copyright 2003,2004 Olly Betts
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -103,10 +103,10 @@ OmExpandWeight::get_weight(const OmExpandBits &bits,
 			   const string &tname) const
 {
     DEBUGCALL(MATCH, Xapian::weight, "OmExpandWeight::get_weight", "[bits], " << tname);
-    double termfreq = (double)bits.termfreq;
+    double termfreq = double(bits.termfreq);
     if (bits.dbsize != dbsize) {
 	if (bits.dbsize > 0 && !use_exact_termfreq) {
-	    termfreq *= (double) dbsize / (double)bits.dbsize;
+	    termfreq *= double(dbsize) / bits.dbsize;
 	    DEBUGLINE(WTCALC, "Approximating termfreq of `" << tname << "': " <<
 		      bits.termfreq << " * " << dbsize << " / " <<
 		      bits.dbsize << " = " << termfreq << " (true value is:" <<
